Move prometheusutils and signalmessenger to pkg
