+++ pkg/prometheusutils/testdata/PrometheusExpositionFormatParser/metrics.txt
